move core logic
