+++ src/helper/ssg/ssg.test.tsx
@@ -4,13 +4,13 @@ import { Hono } from '../../hono'
 import { jsx } from '../../jsx'
 import { poweredBy } from '../../middleware/powered-by'
 import { SSG_DISABLED_RESPONSE, ssgParams, isSSGContext, disableSSG, onlySSG } from './middleware'
-import { fetchRoutesContent, saveContentToFile, toSSG } from './index'
+import { fetchRoutesContent, saveContentToFile, toSSG } from './ssg'
 import type {
   BeforeRequestHook,
   AfterResponseHook,
   AfterGenerateHook,
   FileSystemModule,
-} from './index'
+} from './ssg'
 
 const resolveRoutesContent = async (res: ReturnType<typeof fetchRoutesContent>) => {
   const htmlMap = new Map<string, { content: string | ArrayBuffer; mimeType: string }>()
